pyproject.toml
[project]
name = "pygame-panel"
version = "0.0.3"
authors = [
    { name = "Tomatofu" }
]
description = "An opinionated library for making simple panels in Pygame-CE."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.5.6",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/tomatofoo/pygame-panel"
Issues = "https://github.com/tomatofoo/pygame-panel/issues"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pygame-panel"]

